+++ tests/conftest.py
@@ -1,0 +1,10 @@
+import os.path
+
+import pytest
+
+TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "test-data")
+
+
+@pytest.fixture
+def test_data_dir():
+    return TEST_DATA_DIR